+++ src/NetworkManager.py
@@ -1,5 +1,5 @@
 # Goat - Pico Network Manager library
-# Version 1.1.6
+# Version 1.1.7
 # © (c) 2024-2025 Goat Technologies
 # Description:
 # Provides network management for your device firmware.
@@ -33,7 +33,7 @@ class NetworkManager:
         self.config_file = "network_config.conf"
 
         # Constants
-        self.VERSION = "1.1.6"
+        self.VERSION = "1.1.7"
         self.repo_url = "https://github.com/CodeGoat-dev/Pico-Network-Manager"
 
         # Interface configuration
@@ -486,7 +486,7 @@ class NetworkManager:
             parsed_time[5], 0)  # Subseconds set to 0
 
                 # Create RTC object
-                if not isinstance(self.rtc, object):
+                if not isinstance(self.rtc, machine.RTC):
                     self.rtc = machine.RTC()
 
                 # Set RTC time